# stock-python/contracts/openapi/public_api.yaml
openapi: 3.0.3
info:
  title: Stock API - Public
  description: Public API endpoints for stock trading platform
  version: 1.0.0
  contact:
    name: Stock API Support
servers:
  - url: /api/v1
    description: Public API v1
...
paths:
...
  /auth/verify:
    post:
      summary: Verify email
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, code]
              properties:
                email:
                  type: string
                  format: email
                code:
                  type: string
      responses:
        '200':
          description: Email verified
  /auth/refresh:
    post:
      summary: Refresh access token
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
...
  /signals:
    get:
      summary: List signals
      tags: [signals]
      parameters:
        - name: symbol
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: signal_type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: List of signals
    post:
      summary: Create signal manually
      tags: [signals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stock_id:
                  type: integer
                signal_type:
                  type: string
                entry_price:
                  type: number
                stop_loss:
                  type: number
                take_profit_1:
                  type: number
                take_profit_2:
                  type: number
                take_profit_3:
                  type: number
                reasoning:
                  type: string
      responses:
        '200':
          description: Signal created
...
  /signals/{signal_id}:
    patch:
      summary: Update signal
      tags: [signals]
      parameters:
        - name: signal_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                stop_loss:
                  type: number
                take_profit_1:
                  type: number
                take_profit_2:
                  type: number
                take_profit_3:
                  type: number
      responses:
        '200':
          description: Signal updated
...
  /notifications:
    get:
      summary: Get user notifications
      tags: [notifications]
      parameters:
        - name: skip
          in: query
          schema:
            type: integer
            default: 0
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: unread_only
          in: query
          schema:
            type: boolean
            default: false
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of notifications
...
  /notifications/{notification_id}:
    delete:
      summary: Delete notification
      tags: [notifications]
      parameters:
        - name: notification_id
          in: path
          required: true
          schema:
            type: integer
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Notification deleted

  # Devices
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
...